Added uxtourney tournament system changes.
If the mode is tournament, players will be assigned to slots based on their tournament status and the layout specified in map configuration options. Chat messages will be pushed to the uxtourney chat database. GHost will update the game status in the database when it starts and then again when the game ends.

Stats: stats will be saved in uxtourney_res_ tables rather than the defaults. This includes both the game information and all players, as well as DotA / W3MMD statistics.

Win conditions: CMap also has a m_Conditions string that can specify win conditions for stats. Currently, CStatsDotA uses this and win conditions based on number of tower kills, number of kills, and time limit are allowable.