Draw the game by command or if too many players leave at the beginning.
Draw means that neither team will be recorded as a winner and the game will be ended in sixty seconds.

Also added feature to set the only remaining team as winner (in case the other team leaves). When this happens, the game "soft" ends, meaning stats are locked and autobans don't count but players aren't force kicked.