Add streaming capability.
This allows players to "stream" a game through their own Warcraft III client as if they were an observer. The stream has a delay equal to the time that the streamer begins streaming the game minus the time when the game started. There is also a preset minimum delay to prevent ghosting, which can be set in configuration.

The stream requires an observer slot with a fake player (or a real player). The streamer will see himself/herself in this slot (this is necessary to sync properly).

Actions and chat from the streamer are ignored, so for example pings won't work; the former is to prevent desyncing, while the latter is to prevent ghosting/other abuse.

The streamer can see all chat and also chat send to observers/referees in the game. No other chat is visible. This chat is displayed with the delay, like watching a replay.

Players in the game are also alerted when a streamer begins or stops streaming. Note that this is displayed in all chat, so the streamer will see this message when a time period equal to the delay has elapsed.